+++ src/app/test/page.js
@@ -1,0 +1,13 @@
+export default function Home() {
+    return (
+      <div className="grid max-w-4xl mx-auto grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20 font-[family-name:var(--font-geist-sans)]">
+        <main className="flex flex-col gap-8 row-start-2 items-center sm:items-center w-full">
+  
+          <div className="flex gap-4 items-center flex-col sm:flex-col w-full">
+            <div>
+            </div>
+          </div>
+        </main>
+      </div>
+    );
+  }
